refactor: make a function for formatting response for a single endpoint
